added backend with candidate modules

+++ backend/teraHire/src/main/java/com/octopus/teraHire/service/UserService.java
@@ -1,52 +1,74 @@
 package com.octopus.teraHire.service;
 
+import com.octopus.teraHire.dbutil.DBUtil;
 import com.octopus.teraHire.exception.ResourceNotFoundException;
 import com.octopus.teraHire.exception.UserExistsException;
 import com.octopus.teraHire.model.User;
 import com.octopus.teraHire.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService implements UserInterface{
+    Connection connection;
+    int flag = 0;
     private UserRepository userRepository;
     private UserExistsException userExistsException;
     public UserService(UserRepository userRepository){
         this.userRepository = userRepository;
-
     }
 
+    public UserService() throws SQLException{
+
+        connection = DBUtil.getConnection();
+    }
+
+    //checking existing email
     public boolean isUserEmailExists(String emailId){
         return userRepository.existsByEmail(emailId);
     }
 
+    //setDate
     public LocalDateTime getDate(){
         //DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
         LocalDateTime now = LocalDateTime.now();
         return now;
     }
+
+    //Add users
+
+
     @Override
     public ResponseEntity addNewUser(User user) {
+
+        //password hashing
+        BCryptPasswordEncoder bcrypt = new BCryptPasswordEncoder();
+        String encryptedPwd = bcrypt.encode(user.getPassword());
+        user.setPassword(encryptedPwd);
 
         if (!isUserEmailExists(user.getEmail())) {
             user.setCreatedDate(getDate());
             user.setModifiedDate(getDate());
-            if (userRepository.save(user) == user) {
-                return new ResponseEntity<User>(user, HttpStatus.OK);
-
-            } else {
-                return new ResponseEntity<>(new UserExistsException("Error occurred").getLocalizedMessage(), HttpStatus.FOUND);
-            }
+            return new ResponseEntity<User>(userRepository.save(user), HttpStatus.OK);
         } else {
 
             return new ResponseEntity<>(new UserExistsException("User already exists").getLocalizedMessage(), HttpStatus.FOUND);
         }
     }
 
+    //UpdateUsers
     @Override
     public ResponseEntity updateNewUser(long id, User userDetails){
         User updateNewUser = userRepository.getReferenceById(id);
@@ -58,21 +80,49 @@ public class UserService implements UserInterface{
 
            return new ResponseEntity<>(userRepository.save(updateNewUser),HttpStatus.OK);
         }else{
-            return new ResponseEntity<>(new ResourceNotFoundException("user not exist with id:" + id).getMessage(),HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(new ResourceNotFoundException("user does not exist with this id:" + id).getMessage(),HttpStatus.NOT_FOUND);
         }
-
-
-
     }
+    //DeleteUser
     @Override
-    public ResponseEntity<User> deleteUserById(long id){
-        userRepository.deleteById(id);
-        return null;
+    public ResponseEntity deleteUserById(long id){
+        if(userRepository.existsById(id)){
+            userRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        else {
+            return new ResponseEntity<>(new ResourceNotFoundException("user not exist with id: " + id).getMessage(),HttpStatus.NOT_FOUND);
+        }
     }
 
+    //ListUsers
     @Override
     public List<User> getUserList(){
         return userRepository.findAll();
+    }
+
+    @Override
+    public int loginValidation(String email, String password) {
+
+        try {
+            PreparedStatement statement = connection.prepareStatement("SELECT * FROM user_table WHERE email = '"+email+"' ");
+            ResultSet rs = statement.executeQuery();
+
+            while (rs.next()){
+                if(rs.getString(6).equals(email) && rs.getString(3).equals(password)){
+                    flag = 1;
+                }
+                else {
+                    System.out.println("Invalid Email/password");
+                    flag = 0;
+                }
+            }
+
+
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+        return flag;
     }
 
 }
